HF-Notebook/requirements.txt
# Core dependencies for Oura Ring data analysis notebook

# Environment variable management
python-dotenv>=0.19.0

# Data manipulation and analysis
pandas>=1.3.0
numpy>=1.21.0

# Visualization
plotly>=5.0.0
matplotlib>=3.4.0

# Date/time handling
pytz>=2021.1

# HTTP requests for Oura API
requests>=2.26.0

# Statistical analysis (for Bland-Altman plots)
scipy>=1.7.0

# Jupyter notebook support
ipywidgets>=7.6.0
IPython>=7.25.0

# Excel file support
openpyxl>=3.0.7
xlrd>=2.0.1
